doc: update-ref: discuss symbolic links
Add a paragraph which just emphasizes that the command without any
options does not support refs in the final arguments.  This is
clear already from the names `<new-oid>` and `<old-oid>` but the right
balance of redundancy makes documentation robust to stray
interpretation.

This is also a good place to mention why `--stdin` has those `symref-*`
commands.

Signed-off-by: Kristoffer Haugsbakk <code@khaugsbakk.name>
Signed-off-by: Taylor Blau <me@ttaylorr.com>